+++ jablib/src/test/java/org/jabref/logic/importer/fetcher/LibraryOfCongressTest.java
@@ -2,20 +2,19 @@ package org.jabref.logic.importer.fetcher;
 
 import java.util.Optional;
 
-import org.jabref.logic.importer.FetcherClientException;
 import org.jabref.logic.importer.FetcherException;
 import org.jabref.logic.importer.ImportFormatPreferences;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.BibEntryPreferences;
 import org.jabref.model.entry.field.StandardField;
 import org.jabref.model.entry.field.UnknownField;
+import org.jabref.model.entry.types.UnknownEntryType;
 import org.jabref.testutils.category.FetcherTest;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
@@ -35,17 +34,19 @@ class LibraryOfCongressTest {
 
     @Test
     void performSearchById() throws FetcherException {
-        BibEntry expected = new BibEntry()
+        BibEntry expected = new BibEntry();
+        expected.setType(new UnknownEntryType("text"));
+        expected
                 .withField(StandardField.ADDRESS, "mau, Burlington, MA")
                 .withField(StandardField.AUTHOR, "West, Matthew")
                 .withField(StandardField.DATE, "2011")
-                .withField(StandardField.ISBN, "0123751063 (pbk.)")
+                .withField(StandardField.ISBN, "0123751063")
                 .withField(new UnknownField("issuance"), "monographic")
                 .withField(StandardField.KEYWORDS, "Database design, Data structures (Computer science)")
                 .withField(StandardField.LANGUAGE, "eng")
                 .withField(new UnknownField("lccn"), "2010045158")
                 .withField(StandardField.NOTE, "Matthew West., Includes index.")
-                .withField(new UnknownField("oclc"), "ocn665135773")
+                .withField(new UnknownField("oclc"), "665135773")
                 .withField(new UnknownField("source"), "aacr")
                 .withField(StandardField.TITLE, "Developing high quality data models")
                 .withField(StandardField.YEAR, "2011");
@@ -59,7 +60,7 @@ class LibraryOfCongressTest {
     }
 
     @Test
-    void performSearchByInvalidId() {
-        assertThrows(FetcherClientException.class, () -> fetcher.performSearchById("xxx"));
+    void performSearchByInvalidId() throws FetcherException {
+        assertEquals(Optional.empty(), fetcher.performSearchById("xxx"));
     }
 }